Fix: 싱글 플레이 랭킹 조회 방식 수정

+++ backend/src/main/java/com/vita/backend/character/data/response/detail/RequesterGameSingleRankingDetail.java
@@ -11,6 +11,6 @@ public record RequesterGameSingleRankingDetail(
 	@JsonProperty("ranking")
 	Long ranking,
 	@JsonProperty("score")
-	Double score
+	Long score
 ) {
 }
